シネマシティズンクラスを60以上のみにしていたが、シネマシティズンである人を処理するように変更。
